[!!!][TASK] Add initialize file after first ddev start into project
The way to save an initialize file into the docker container are not
working after ddev update to version 0.19.0. Now the file would be
located in the build-folder.

If you need fresh a fresh installation of your project after the first
start of ddev, remove the initialize file and run `ddev remove -R` to
delete all container and start with a fresh installation.